:recycle: refactor!: reorganize the tag structure
