+++ lib/tag/create.js
@@ -1,9 +1,9 @@
 import fs from "node:fs";
 import path from "node:path";
 
-import { config } from "./config.js";
-import { run } from "./utils.js";
-import { computeVersion } from "./version/compute.js";
+import { config } from "../config.js";
+import { run } from "../utils.js";
+import { computeVersion } from "../version/compute.js";
 
 /* ===========================
  * Core API (Programmatic)
