+++ package.json
@@ -11,7 +11,7 @@
     "rs-changelog": "bin/changelog.js",
     "rs-generate-release-notes": "bin/generate-release-notes.js",
     "rs-dry-run": "bin/dry-run.js",
-    "rs-create-tag": "bin/create-tag.js"
+    "rs-tag": "bin/tag.js"
   },
   "scripts": {
     "lint": "eslint bin/",
@@ -21,7 +21,7 @@
     "version:compute": "node bin/version.js compute",
     "changelog": "node bin/changelog.js generate",
     "release-notes": "node bin/generate-release-notes.js",
-    "create-tag": "node bin/create-tag.js",
+    "tag:create": "node bin/tag.js create",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "engines": {
